+++ blank/page-custom.php
@@ -11,7 +11,7 @@ For more info: http://codex.wordpress.org/Page_Templates
     <div class="content clearfix">
 
         <!-- section -->
-        <section role="main">	           
+        <section class="main" role="main">	           
 
             <?php if (have_posts()) : while (have_posts()) : the_post(); ?>