+++ blank/page.php
@@ -4,7 +4,7 @@
     <div class="content clearfix">
 
         <!-- section -->
-        <section role="main">	           
+        <section class="main" role="main">	           
 
             <?php if (have_posts()) : while (have_posts()) : the_post(); ?>