+++ blank/single.php
@@ -4,13 +4,11 @@
     <div class="content clearfix">
 
         <!-- section -->
-        <section role="main">	
+        <section class="main" role="main">	
 
             <?php if (have_posts()) : while (have_posts()) : the_post(); ?>
 
-            <?php            
-            get_template_part( 'format-default', get_post_format() );            
-            ?>    
+            <?php get_template_part( 'format-default', get_post_format() ); ?>    
 
             <?php endwhile; ?>